refactor: 공통 설정 파일 패키지 이동
