+++ perfume-api/src/main/java/io/perfume/api/common/advices/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package io.perfume.api.advice;
+package io.perfume.api.common.advices;
 
 import io.perfume.api.base.CustomHttpException;
 import io.perfume.api.base.LogLevel;
